+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [4.0.0-beta.7] - 2026-07-06
+## [4.0.0-beta.8] - 2026-07-06
 
 ### Fixed